app/requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
sqlalchemy==1.4.41
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dotenv==0.21.0
databases
asyncpg==0.28.0